Add Google Test in project. Create from Matrix Object Module.
